+++ lib/src/eval/compiler/statement/try.dart
@@ -2,9 +2,11 @@ import 'package:analyzer/dart/ast/ast.dart';
 import 'package:dart_eval/dart_eval_bridge.dart';
 import 'package:dart_eval/src/eval/compiler/context.dart';
 import 'package:dart_eval/src/eval/compiler/errors.dart';
+import 'package:dart_eval/src/eval/compiler/macros/branch.dart';
 import 'package:dart_eval/src/eval/compiler/statement/statement.dart';
 import 'package:dart_eval/src/eval/compiler/type.dart';
 import 'package:dart_eval/src/eval/runtime/runtime.dart';
+import 'package:dart_eval/src/eval/shared/types.dart';
 
 import '../variable.dart';
 
@@ -20,21 +22,16 @@ StatementInfo compileTryStatement(TryStatement s, CompilerContext ctx, AlwaysRet
   if (s.catchClauses.isEmpty) {
     throw CompileError('Try statements must have at least one catch clause');
   }
-  if (s.catchClauses.length > 1) {
-    throw CompileError('Multiple catch clauses are not supported yet');
-  }
 
   final jumpOver = ctx.pushOp(JumpConstant.make(-1), JumpConstant.LEN);
   ctx.rewriteOp(tryOp, Try.make(ctx.out.length), 0);
 
-  final catchClause = s.catchClauses.first;
-
   ctx.beginAllocScope();
   ctx.pushOp(PushReturnValue.make(), PushReturnValue.LEN);
   final v = Variable.alloc(ctx, EvalTypes.dynamicType);
-  ctx.setLocal(catchClause.exceptionParameter!.name.value() as String, v);
-  final catchInfo = compileStatement(catchClause.body, expectedReturnType, ctx);
+  final catchInfo = _compileCatchClause(ctx, s.catchClauses, 0, v, expectedReturnType);
   ctx.endAllocScope();
+
   ctx.rewriteOp(jumpOver, JumpConstant.make(ctx.out.length), 0);
 
   if (s.finallyBlock != null) {
@@ -43,3 +40,34 @@ StatementInfo compileTryStatement(TryStatement s, CompilerContext ctx, AlwaysRet
 
   return bodyInfo | catchInfo.copyWith(willAlwaysThrow: false);
 }
+
+StatementInfo _compileCatchClause(CompilerContext ctx, List<CatchClause> clauses, int index, Variable exceptionVar,
+    AlwaysReturnType? expectedReturnType) {
+  final catchClause = clauses[index];
+  final exceptionType = catchClause.exceptionType;
+  if (exceptionType == null) {
+    ctx.setLocal(catchClause.exceptionParameter!.name.value() as String, exceptionVar);
+    return compileStatement(catchClause.body, expectedReturnType, ctx);
+  }
+  final slot = TypeRef.fromAnnotation(ctx, ctx.library, exceptionType);
+  return macroBranch(
+    ctx,
+    expectedReturnType,
+    condition: (_ctx) {
+      ctx.pushOp(IsType.make(exceptionVar.scopeFrameOffset, runtimeTypeMap[slot] ?? ctx.typeRefIndexMap[slot]!, false),
+          IsType.LEN);
+      return Variable.alloc(ctx, EvalTypes.boolType.copyWith(boxed: false));
+    },
+    thenBranch: (_ctx, _expectedReturnType) {
+      ctx.setLocal(catchClause.exceptionParameter!.name.value() as String, exceptionVar.copyWith(type: slot));
+      return compileStatement(catchClause.body, expectedReturnType, ctx);
+    },
+    elseBranch: clauses.length <= index + 1
+        ? null
+        : (ctx, expectedReturnType) {
+            return _compileCatchClause(ctx, clauses, index + 1, exceptionVar, expectedReturnType);
+          },
+  );
+}
+
+///